stopped using VLAs to support older harware

+++ src/libs/file.h
@@ -2,6 +2,7 @@
 #define FILE_H
 
 #include <stdbool.h>
+#include <stddef.h>
 #include <sys/stat.h>
 
 #define CACHE_FILE "weather_cache.json"
@@ -10,7 +11,7 @@
 #define MAX_TIME_SINCE_LAST_FETCH 900
 
 int write_cache(const char *city, const char *json);
-int read_cache(const char *city, char *output, int size);
+int read_cache(const char *city, char *output, size_t size);
 
 void getFileCreationTime();
 bool time_for_new_data();
